Updated to auto set DH parameters.
Updated to use well known Diffie-Hellman (DH) parameters that have built-in support in OpenSSL.
This means the DH parameters will be selected to be consistent with the size of the key associated with the server's certificate.
If there is no certificate (e.g. for PSK ciphersuites), then it it will be consistent with the size of the negotiated symmetric cipher key.